[kernels][tests] Use fresh_knobs when touching triton.knobs
