+++ python/triton_kernels/tests/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+from triton._internal_testing import _fresh_knobs_impl
 
 
 def pytest_addoption(parser):
@@ -8,3 +9,12 @@ def pytest_addoption(parser):
 @pytest.fixture
 def device(request):
     return request.config.getoption("--device")
+
+
+@pytest.fixture
+def fresh_knobs(monkeypatch):
+    fresh_function, reset_function = _fresh_knobs_impl(monkeypatch)
+    try:
+        yield fresh_function()
+    finally:
+        reset_function()
